requirements.txt
transformers
celery
fastapi
jinja2
websockets
pydantic
google-generativeai
asyncio
redis
anthropic
python-dotenv
openai